+++ src/components/ui/MobileBtn/MobileBtn.tsx
@@ -1,0 +1,17 @@
+"use client";
+import React from "react";
+
+type MobileBtnProps = {
+  isMobileMenuShown: boolean;
+  handleClick: () => void;
+};
+
+const MobileBtn = ({ isMobileMenuShown, handleClick }: MobileBtnProps) => {
+  return (
+    <button className="w-[24px] h-[24px] block lg:hidden" onClick={handleClick}>
+      {!isMobileMenuShown ? "|||" : "X"}
+    </button>
+  );
+};
+
+export default MobileBtn;
