+++ src/components/ui/MobileMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MobileMenu";
