+++ src/components/ui/NavBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NavBar";
